src/main/java/uk/thepragmaticdev/security/UserService.java
package uk.thepragmaticdev.security;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import uk.thepragmaticdev.account.AccountRepository;

@Service
public class UserService implements UserDetailsService {

  private final AccountRepository accountRepository;

  @Autowired
  public UserService(AccountRepository accountRepository) {
    this.accountRepository = accountRepository;
  }

  @Override
  public UserDetails loadUserByUsername(String username) {
    final var account = accountRepository.findByUsername(username)
        .orElseThrow(() -> new UsernameNotFoundException(username));

    return org.springframework.security.core.userdetails.User//
        .withUsername(username)//
        .password(account.getPassword())//
        .authorities(account.getRoles())//
        .accountExpired(false)//
        .accountLocked(false)//
        .credentialsExpired(false)//
        .disabled(false)//
        .build();
  }
}